% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MH_Functions.R
\name{ChromAO}
\alias{ChromAO}
\title{ChromAO}
\usage{
ChromAO(vcf, chromlist = NULL, windowSize = NULL)
}
\arguments{
\item{vcf}{A vcf file}

\item{chromlist}{A vector specifying particular chromosomes}

\item{windowSize}{Specify window size to calculate number of SNPs}
}
\value{
A histogram of of Depth FIELD
}
\description{
ChromAO
}
