% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MH_Functions.R
\name{ChromDP}
\alias{ChromDP}
\title{ChromQual}
\usage{
ChromDP(vcf, chromlist = NULL, windowSize = NULL, binwidth = NULL)
}
\arguments{
\item{vcf}{A vcf file}

\item{chromlist}{A vector specifying particular chromosomes}

\item{windowSize}{Specify window size to calculate number of SNPs}

\item{binwidth}{Specify bindwidth for histogram plot}
}
\value{
A histogram of of Depth FIELD
}
\description{
ChromQual
}
\examples{
ChromQuality(vcf = "General.vcf", chromlist = c("Chr01","Chr02"),windowsize=1e+05,binwidth=10)
}
