% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MH_Functions.R
\name{ChromQual}
\alias{ChromQual}
\title{ChromQual}
\usage{
ChromQual(
  vcf,
  chromlist = NULL,
  windowSize = NULL,
  binwidth = NULL,
  FIELD = NULL
)
}
\arguments{
\item{vcf}{A vcf file}

\item{chromlist}{A vector specifying particular chromosomes}

\item{windowSize}{Specify window size to calculate number of SNPs}

\item{binwidth}{Specify bindwidth for histogram plot}

\item{FIELD}{Specify info field to plot}
}
\value{
A histogram of of FIELD
}
\description{
ChromQual
}
\examples{
ChromQuality(vcf = "General.vcf", chromlist = c("Chr1", "Chr2")), windowSize = 1e+06, scalar = 0.1, ncol = 2,HighLimQuality = 6000,  binwidth1 = 100, binwidth2 =1, DPBINS=10, p1=TRUE, p2=FALSE, p3=TRUE, p4=TRUE, p5=FALSE, p6=TRUE)
}
