% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MH_Functions.R
\name{Correlation}
\alias{Correlation}
\title{Correlation}
\usage{
Correlation(
  vcffile = NULL,
  chromlist = NULL,
  p1 = NULL,
  p2 = NULL,
  p3 = NULL,
  p4 = NULL,
  p5 = TRUE
)
}
\arguments{
\item{vcffile}{A vcf file}

\item{chromlist}{A vector of chromosomes/contigs as specified in VCF file}

\item{p1}{A boolean Value either TRUE or FALSE for plotting Correlation matrix and Correlation Tables}

\item{p2}{A boolean Value either TRUE or FALSE for plotting Correlation matrix and Correlation Tables}

\item{p3}{A boolean Value either TRUE or FALSE for plotting Correlation matrix and Correlation Tables}

\item{p4}{A boolean Value either TRUE or FALSE for plotting Correlation matrix and Correlation Tables}

\item{p5}{A boolean Value either TRUE or FALSE}
}
\description{
Returns a Correlation Matrix for info fields AC, DP, DPB, QA, RO, AO and most importantly QUAL
}
