% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MH_Functions.R
\name{FacetChromAO}
\alias{FacetChromAO}
\title{FacetChromAO}
\usage{
FacetChromAO(vcf, chromlist = NULL, windowSize = NULL, ncol = NULL)
}
\arguments{
\item{vcf}{A vcf file please please}

\item{chromlist}{A vector specifying particular chromosomes}

\item{windowSize}{Specify window size to calculate number of SNPs}

\item{ncol}{An integer representing the number of Chromosomes in your set list}
}
\description{
FacetChromAO
}
