% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MH_Functions.R
\name{FacetChromDP}
\alias{FacetChromDP}
\title{FacetChromDP}
\usage{
FacetChromDP(vcf, chromlist = NULL, windowSize = NULL, ncol = NULL, bins = NULL)
}
\arguments{
\item{vcf}{A vcf file please}

\item{chromlist}{A vector specifying particular chromosomes}

\item{windowSize}{Specify window size to calculate number of SNPs}

\item{ncol}{An integer representing the number of Chromosomes in your set list}
  
\item{bins}{Specify bin width of histogram having a better more accurate representation}
}
\description{
FacetChromDP
}
